% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Rmd.R
\name{R2Rmd}
\alias{R2Rmd}
\title{Function to convert R file into Rmd file with chunks}
\usage{
R2Rmd(fileName, outFormat = "unknown", rmessage = FALSE,
  rwarning = FALSE, textTitle = NULL, Author = NULL, Date = NULL,
  saveFileName = NULL, plotEcho = TRUE)
}
\arguments{
\item{fileName}{The R file name which you want to refer.}

\item{outFormat}{The output format for the subtitle (chapter title). The default is "unknown", and
this will result in the simplest format (the number of "#" will always be 3 in Rmd file).
The other is outFormat = 1, which returns the number of "#" in Rmd file as below.

(the number of "#" in Rmd file) = 8 - (the number of "#" in R file (if > 2))

(the number of "#" in Rmd file) = 0 (if (the number of "#" in R file) <= 2)}

\item{rmessage}{Show message for chunks in the Rmd file or not.}

\item{rwarning}{Show warning for chunks in the Rmd file or not.}

\item{textTitle}{The title of the Rmd file.

If textTitle = NULL, the file name will be used.}

\item{Author}{The author of the Rmd file.}

\item{Date}{The date of the Rmd file.}

\item{saveFileName}{The name of the Rmd file to be saved.

If saveFileName = NULL,  the file name will be used. (only file extension will change)}

\item{plotEcho}{Show the code for plotting in chunks or not.}
}
\value{
Nothing. Coverted ".Rmd" file will be created.
}
\description{
Function to convert R file into Rmd file with chunks
}
