% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd2R.R
\name{Rmd2R}
\alias{Rmd2R}
\title{Function to convert Rmd file into R file}
\usage{
Rmd2R(fileName, outFormat = "unknown", Numbering = TRUE,
  remainExplanation = TRUE, textTitle = NULL, addTitle = TRUE,
  saveFileName = NULL)
}
\arguments{
\item{fileName}{The Rmd file name which you want to refer.}

\item{outFormat}{The output format for the subtitle (chapter title). The default is "unknown", and
this will result in the simplest format (the number of "#" will always be 4 in R file).
The other is outFormat = 1, which returns the number of "#" in Rmd file as below.

(the number of "#" in R file) = 8 - (the number of "#" in Rmd file)}

\item{Numbering}{Add chapter numbers for each subtitle (chapter title) or not.}

\item{remainExplanation}{Remain explanations in the original Rmd file or remove those.}

\item{textTitle}{The title of the R file.

If textTitle = NULL, the file name will be used.}

\item{addTitle}{Add title part in the R file or not.}

\item{saveFileName}{The name of the R file to be saved.

If saveFileName = NULL,  the file name will be used. (only file extension will change)}
}
\value{
Nothing. Coverted ".R" file will be created.
}
\description{
Function to convert Rmd file into R file
}
